Update of the functionalities
I changed the proposal in order to be in accordance with the technologies we will be using. The streamming of music doesn't seem so difficult to do using Mopidy, so it was putted in the v1. Moreover, the Mopidy communicates with a smartphone/tablet/PC using a HTTP/MPD client. Hence, I put this kind of interaction in the document and  deleted the proposal of doing the communication over Bluetooth. After all, the server doesn't support this. Finally, I putted Alexa in the v2. Despite thinking this is the more exciting part of the project, I strongly feel we should try to do the basic first and then try to combine the two technologies.